+++ test/TestVector.cpp
@@ -11,7 +11,7 @@ static int roundDown(const double &num) {
 
 void testVector() {
     Vector<double, 5> x;
-    for (int i = 0; i < 5; i++) assert(x[i] == 0);
+    for (size_t i = 0; i < 5; i++) assert(x[i] == 0);
     Vector<double, 5> y;
     x[0] = 2;
     y[0] = 4;
@@ -29,10 +29,10 @@ void testVector() {
 
     Vector<double, 5> test{1.1, -4.5, 3, 4, 0};
     Vector<int, 5> test_floor = test.applyFunc<int>(&roundDown);
-    for (int i = 0; i < 5; i++) assert(test_floor[i] == (int) test[i]);
+    for (size_t i = 0; i < 5; i++) assert(test_floor[i] == (int) test[i]);
 
     Vector<int, 5> test_sq_int = test.applyFunc<int>([](const double &d) { return (int) (d * d); });
-    for (int i = 0; i < 5; i++) assert(test_sq_int[i] == (int) (test[i] * test[i]));
+    for (size_t i = 0; i < 5; i++) assert(test_sq_int[i] == (int) (test[i] * test[i]));
 
     Vector<ExprPtr, 3> expr{Variable::make("x"),
                             Variable::make("y"),
@@ -41,7 +41,7 @@ void testVector() {
                                              {"y", 1},
                                              {"z", 2}};
     Vector<double, 3> expr_sub = expr.applyFunc<double>([&subs](const ExprPtr &e) { return e->evaluate(subs); });
-    for (int i = 0; i < 3; i++) assert(expr_sub[i] == i);
+    for (size_t i = 0; i < 3; i++) assert(expr_sub[i] == i);
 
     std::cout << "Passed All Tests for Vector!" << std::endl;
 }
